Import and normalize existing presentations
ACM has many seminars, but they are fragmented, and differ in markup
style and location. This commit brings them into one place, and -- for
the most part -- unifies their markup style and naming conventions to
align with future style guidelines.

Note that the new markup works well (generally) with reveal.js'
automatic conversion to slides. Because we can now use vertical slide
arrangement, it may be worthwhile to adjust certain presentations
with that in mind.