Make URL auto-fetching Robert-proof
Fetching the markdown file properly based on the search string used to
be somewhat insecure (almost-arbitrary access to markdown files), and
wouldn't work if you added a slash after the presentation name.

I've fixed both of those things by extracting only what's before the
first slash (if any) in the search string, and using that to pick the
presentation.